feat: update koii data ingestion scriopt
